+++ src/main/java/lotto/utils/StringUtils.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.utils;
 
 import java.util.Arrays;
 import java.util.List;
