+++ src/test/java/lotto/utils/StringUtilsTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.utils;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
